fix(protocol): multiply vars with "*" operator instead of ".mul"
chore(protocol): fix typo "denumerator"
chore(protocol): use "amount" instead of "value" in test case descriptors
docs(protocol): enhance natspec for "getRepayAmount" function
docs(protocol): simplify natspec for "newDebtCeiling" return var in "setDebtCeiling" function
refactor(protocol): delete redundant check for "liquidationIncentive" in "getRepayAmount"
refactor(protocol): reorder imports in "BalanceSheetV1.sol"
refactor(protocol): reorder arguments in "getRepayAmount"
refactor(protocol): rename "getRepayBondAmount" to "getRepayAmount"
style(protocol): sort imports in some typescript files
test(protocol): delete stale tests re "liquidationIncentive" for "getRepayAmount" function
