+++ packages/protocol/test/unit/balanceSheet/BalanceSheet.behavior.ts
@@ -1,5 +1,11 @@
+import shouldBehaveLikeUpgradeableProxy from "./deployment/proxy";
 import shouldBehaveLikeBorrow from "./effects/borrow";
 import shouldBehaveLikeDepositCollateral from "./effects/depositCollateral";
+import shouldBehaveLikeLiquidateBorrow from "./effects/liquidateBorrow";
+import shouldBehaveLikeRepayBorrow from "./effects/repayBorrow";
+import shouldBehaveLikeRepayBorrowBehalf from "./effects/repayBorrowBehalf";
+import shouldBehaveLikeSetOracle from "./effects/setOracle";
+import shouldBehaveLikeWithdrawCollateral from "./effects/withdrawCollateral";
 import shouldBehaveLikeFintrollerGetter from "./view/fintroller";
 import shouldBehaveLikeGetBondList from "./view/getBondList";
 import shouldBehaveLikeGetCollateralAmount from "./view/getCollateralAmount";
@@ -7,15 +13,9 @@ import shouldBehaveLikeGetCollateralList from "./view/getCollateralList";
 import shouldBehaveLikeGetCurrentAccountLiquidity from "./view/getCurrentAccountLiquidity";
 import shouldBehaveLikeGetDebtAmount from "./view/getDebtAmount";
 import shouldBehaveLikeGetHypotheticalAccountLiquidity from "./view/getHypotheticalAccountLiquidity";
-import shouldBehaveLikeGetRepayBondAmount from "./view/getRepayBondAmount";
+import shouldBehaveLikeGetRepayAmount from "./view/getRepayAmount";
 import shouldBehaveLikeGetSeizableCollateralAmount from "./view/getSeizableCollateralAmount";
-import shouldBehaveLikeLiquidateBorrow from "./effects/liquidateBorrow";
 import shouldBehaveLikeOracleGetter from "./view/oracle";
-import shouldBehaveLikeRepayBorrow from "./effects/repayBorrow";
-import shouldBehaveLikeRepayBorrowBehalf from "./effects/repayBorrowBehalf";
-import shouldBehaveLikeSetOracle from "./effects/setOracle";
-import shouldBehaveLikeUpgradeableProxy from "./deployment/proxy";
-import shouldBehaveLikeWithdrawCollateral from "./effects/withdrawCollateral";
 
 export function shouldBehaveLikeBalanceSheet(): void {
   describe("Deployment", function () {
@@ -51,8 +51,8 @@ export function shouldBehaveLikeBalanceSheet(): void {
       shouldBehaveLikeGetHypotheticalAccountLiquidity();
     });
 
-    describe("getRepayBondAmount", function () {
-      shouldBehaveLikeGetRepayBondAmount();
+    describe("getRepayAmount", function () {
+      shouldBehaveLikeGetRepayAmount();
     });
 
     describe("getSeizableCollateralAmount", function () {
